Updated openstack/openstack Project: openstack/requirements  b9871b91f956db16129a37efc8284f25f1d11f08
hacking: bump minimal version to >= 0.11.0

This is needed to get access to off_by_default decorator that allows to
add checks that are not enabled by default. This is needed when hacking
checks are shared between repositories, and an introduction of a check
in one repository may break a gate for another one. This is a pattern in
Neutron.

The patch that added off_by_default support:
If188f85e8093974802bd2bacc2a72b2ae35ee5a1

Change-Id: I883223962426ed8e9cdf4e304dfed17ef0946c8a