Updated openstack/openstack Project: openstack/trove  4f888ddfa38e1002f12a2b3ef24b833a32f3f1d4
pkg: replace commands module with subprocess

The commands is deprecated since Python 2.6 and was removed from
Python 3. This change replaces it with subprocess. By the way, it
also changes the code to avoid a shell and so indirectly makes the
code safer (shell injections are no more possibles).

tox: run test_backupagent and test_pkg on Python 3.4.

Partially implements: blueprint trove-python3
Change-Id: I3818fa1498819cc77a3da42c12245c50b96cad59