Updated openstack/openstack Project: openstack/api-wg  f543bed8a06c842dd5dc70a5d2813f9f50b2689d
Description of how to use etags to avoid lost updates

This new document describes etags and how they can be used to
avoid the lost update problem. To avoid confusion it contains an
introduction that explains that it does not cover all of the many
use and corner cases for etags.

Change-Id: Iacb7cb7cddef32b3cabd371255ae8443ce5beb80