Updated openstack/openstack Project: openstack/horizon  caa5e910598707d2cfc56faedb6dff97c574e50c
Fix Keystone version suffix when Keystone webpath is present

In this patch the URLs that are coming from service catalog are fixed
to contain the proper version.

Update requirements.txt in the same commit, because otherwise
integration tests won't pass for OpenStack Proposal Bot commit which
should update DOA version.

Closes-Bug: #1585682
Depends-On: Iea9b8e8378e6c5fb4c60df0073968d8caf7fbc5e
Change-Id: Icebfc291ec2b06ed84934c75cfd8c9d91cb2a895