Updated openstack/openstack Project: openstack/requirements  2d2d21d01223c2ef7d424d4f23aec7bb5811801a
Raise mock requirement to 2.0

Mock now has assert_called and assert_called_once checks. We had
previously needed to watch out for these and replace them with
things like assertEqual(1, mocked.call_count) if a new patch
introduced them.

This was complicated by the fact that the gate has been using 2.0
for some time as it pulled the latest version, so it wouldn't catch
if there was an invalid call introduced. If a user was using a
previously installed version it would fail for them locally.

Rather than having conditional checks or needing to watch for these
incompatibilities in reviews, I propose we just make 2.0 the default.

Change-Id: If8cb4e476ad56d3ce6b01215fdaa2f3aa8a354c4