Updated openstack/openstack Project: openstack-infra/project-config  670bb7a5560e0f1b85b7496a7acaba1fd2ac1f8f
Add Non-voting job for nodepool py34

Add py34 non-voting job for nodepool. With python3 fixes being
proposed to nodepool, it is important that there is at least
minimal reporting on success/failure of the py34 checking. As
soon as the test suite can run this can be converted to a voting
job.

Change-Id: I4fc17b67338eb3f36b550fa9f519d439c0d37022