Updated openstack/openstack Project: openstack-infra/project-config  2c27af3ce7f881eba3453bf2f71cc806562ee40a
Add Ops repo to OpenStack-Ansible

At the Austin summit it was agreed that OpenStack-Ansible should
have a repository to facilitate the sharing of various scripts/tips
which may be useful to operators or developers when using
OpenStack-Ansible.

This patch creates the repository for the project and sets up the
basic jobs for it.

Needed-By: Ib6cd1afcdee502a0819c1afb7b27f27f8f96b381
Change-Id: I96a582d73871c9f479619e5492764919c424e7a1