Updated openstack/openstack Project: openstack-infra/release-tools  dbcef7f9c14f18fbd7519edd46c7835d285ee2b0
update README for the script to expire old bug reports

With commit 623587c76de9deb595404385fbe9a99b58e8e28d a script to expire
old bug reports got introduced. This change updates the docs in the
"README.rst" to reflect that.
It also makes the script executable (chown +x) so that it is
consistent with the other scripts and in this folder.

Change-Id: If6e1d7f79731549e30feeddada0a561d609d56dd