Updated openstack/openstack Project: openstack/trove-specs  e19d73a07a14784a0f04ee080d613d8a2942e75b
Change the order of specs

As we now have four releases of specs (starting with Kilo) when
reviewing the documentation it is more convenient if information were
sorted reverse chronologically. So I've put Newton at the top; when
you go to the website therefore you will see the most recent documents
at the top.

Change-Id: I202a1a33ef6e4f4eb2dbafd79bcd8be468d17440