Updated openstack/openstack Project: openstack-infra/puppet-simpleproxy  7e8b03762f8768ba2fd7fec8ac171ffebafb076e
Create a simpleproxy user

Previously this depended on other things creating a logstash user which
would enable this service to run. There is no logstash user created on
these hosts any longer and it was bad to rely on that side effect of
colocation anyways. Fix this by creating a user just for this service.

Change-Id: I6b0779552d3c47f3bf27b1548599cfc6e9461dcb