Updated openstack/openstack Project: openstack/heat  1e3abae213d9e65b891416a43ff5f5b67ce3b48d
Fix resource loading in resource filtering

We try to be careful when filtering resources from stack, to not load
all of them, but unfortunately it breaks as soon as you have
dependencies with resources that are not returned. To be on the safe
side we need to load them all and filter them afterwards.

Change-Id: I399545a6a4fd77d9707505ec9c4058d384877dec
Closes-Bug: #1585931