Updated openstack/openstack Project: openstack/tripleo-heat-templates  5090ae849b6300dba17fb55c250bc50825f7d21e
Enable proxy header handling for cinder

Cinder is using http_proxy_to_wsgi middleware. This parses the
headers provided by the proxy, and helps us properly use TLS for
keystone discovery. There was an option introduced in this middleware
to have it disabled by default, and this change enables it.

Change-Id: Ia33b3fa04d71eab10effd0b33eb2c194282cd15b