Updated openstack/openstack Project: openstack/nova  f280d2621620fa366e18e6240db85dd83389f4d4
Make Inventory and ResourceProvider objects use the API DB instead

To set the stage for moving this data up, this patch removes the integrated
inventory management from the compute node object (although leaves the
routines for use later). As such, this removes the zeroing behavior of the
ComputeNode legacy resource fields as well, since we can no longer do a
join of the compute_nodes and inventory tables. Thus, we continue to
update both the legacy compute node table columns and the inventory
information in the API database.

Related to blueprint generic-resource-pools

Change-Id: I088b90170463e686cbac49d1d934ae7969f47e6e