Updated openstack/openstack Project: openstack/glance  5cb3a04a82de472948ddb7040d53e8ebc8e7cbb2
Excluded the 'visibility' from protected artifact fields

The visibility field was one of the proptected artifact fields (along
with 'owner', 'state' and modification timestamps) and thus it was
forbidden to explicitly set its value during the artifact creation.

Because of that it was impossible to create a public artifact draft in
a single call, while that may be required.

Also removed a redundant FIXME comment is the same code block since it
is not validRemoved some redundant FIXME comment is the same code
block since it is not valid.

Change-Id: I579405ee2b79b6018a322bb46e34acde6b5b4a4a
Closes-bug: #1585584