Updated openstack/openstack Project: openstack/neutron  d5eac3cfe0f64a792eea0b25aeba342b747bbbd7
Allow the service plugin to import the extension

I hit this circular import which made it so that I couldn't import the
extension from db.py within the service plugin because the extension
imports the plugin.  This seems backwards to me.  The extension is the
interface and should pretty much stand alone shouldn't it?  The
service plugin is the implementation and should be free to import it
at will.

This patch fixes the circular import issue.

Change-Id: I3b84496bf65578b02b43bbffb227db9bb9879288
Partially-implements: bp/routed-networks