Updated openstack/openstack Project: openstack/nova  3a70a7617041ff4b39295dbfa724f19112c0f19f
Fix "KeyError: u'instance_id'" in string format operation

Commit 56ac4418 removed the usage of mox in favor of mock. By mistake it
introduced the mocking of the kwarg "instance_uuid" which isn't part of
the format string of the InstanceNotFound exception. The error looks
like this:

    Exception in string format operation
    Traceback (most recent call last):
      File "nova/exception.py", line 140, in __init__
        message = self.msg_fmt % kwargs
    KeyError: u'instance_id'

This patch changes the mock to use "instance_id", which is the correct
kwarg for the format string of the InstanceNotFound exception. It also
uses the correct "uuids.instance" for this mock.

Change-Id: I6745075e186ad150a8ff1765dbece267e5cc2c5b