Updated openstack/openstack Project: openstack/nova  d77807bfc2ac252a2a4574bd2478d0a416d999e3
VMware: set service status based on vc connection

When the vcenter (vc) session is disconnected, say,
if the management service on vcenter is down, then
the nova-compute service status does not reflect that.

This patch sets disabled=True in nova.objects.Service
when the vcenter session is broken.
When vcenter service is accessible again, then oslo_vmware
takes care of recreating a new session.

Closes-Bug: #1316928
Change-Id: I73a34eb6e0ca32d03e54d12a5e066b2ed4f19a61