Updated openstack/openstack Project: openstack/cinder  1b9d0c98fc7f83cee850ef1484760328ebb12bc3
Volume manage should parse volume size as float

The function manage_existing_object_get_size casts the volume size as
an int, which results in returning the floor of the volume size instead
of the ceiling. It should cast it as a float, which will correctly
result in the ceiling.

Change-Id: I3c8134964e3d8709657fb8bf49d8ba2f9d673149
Closes-Bug: #1584567