Updated openstack/openstack Project: openstack/zaqar  dfa5493a734847d909acd8152aca8af23a8840bb
Improve the log to avoid unexpected ERROR

When notification pipeline and pooling are both enabled, the method
'post' of NotifierDriver will be called two times. One for normal
SubscriptionController and one for pooling.SubscriptionController.
So a check was added to avoid the notifition is sent two times, but
unfortunately an unintentional error was introduced. This patch fixes
it by refactoring the condition checking.

Change-Id: Ie22748e86a50b7d91f24707b0c5a3ab718f005a2