Updated openstack/openstack Project: openstack-infra/subunit2sql  5b8f5dffbe64567a43ba29dfa96b98148e7bdcd7
Fix docs warnings and errors

This commit fixes docs and warnings. But I left some warnings such as
ChangeLog and external image URIs. Because ChangeLog is generated from
git log. And to fix the external image URIs, it needs needs an ugly
workaround that is replacing it to raw html.

Change-Id: Ieecc640fdf5d54d138eeb963ff9ef8cd1c9eaa4a