Updated openstack/openstack Project: openstack-infra/project-config  92a92ad1b127d6c154547f6d9e6815933cacb691
puppet: move puppet4 jobs into check pipeline

Puppet OpenStack group has the desire to run puppet4 jobs as non-voting
in the check pipeline for openstack/puppet-openstack-integration
project.

This patch:
* remove puppet4 periodic jobs.
* create puppet-openstack-integration-jobs-puppet4-check-nv with all
  puppet4 jobs in the check pipeline.
* create puppet-openstack-integration-jobs-puppet4-experimental with
  all puppet4 jobs in the check pipeline.
* Use puppet-openstack-integration-jobs-puppet4-check-nv in
  openstack/puppet-openstack-integration layout, starting from Newton /
  Jewel.
* Use puppet-openstack-integration-jobs-puppet4-experimental-nv in
  Puppet modules layouts that use puppet-openstack-integration-jobs-all.

This is an iteration for us to move to Puppet4 gating.
We started the work with experimental jobs, then periodic jobs and now
check jobs.

Once we stabilize puppet4 jobs, we'll add them in the gate pipeline.

Change-Id: Ie2214d9a46435b7073640974c13709f55760897c