Updated openstack/openstack Project: openstack/openstack-manuals  5933bcded059ce85b9ed93d125fcc54b8318f4e6
Update the entry page with info on SDKs

The current doc talks about "SDK" but is actually then documenting
the python-*client libraries. Since there are a few different choices
these days, including the actual SDK, we should be a bit clearer in
describing what is available and what the differences are.

Note: describing the difference between openstacksdk and shade in a way
that is comprehensive, concise and clear about which things each is
focused on is _hard_ ... further work on verbage here is a really good
idea.

Change-Id: Ie89f704e49dd38709c91c9e5c5cedc9cc98033d7