Updated openstack/openstack Project: openstack/openstack-manuals  a6ba35d2868b0f9d3e81948c7f938549e759d57e
Rework entrypoint to authentication section

Since the compute auth doc was reworked to be an up to date way to get a
novaclient Client object, make the docs clear on that point, and have
the page that gets you there talk about envvars and clouds.yaml files.

Next step will be adding SDK and REST pages.

Change-Id: Idbeec5d39ea26708516333be6f6c14ab3d326698