Updated openstack/openstack Project: openstack/zaqar  86419a6883bd3b089040894c5b9370abea4de9c3
Init commit to add API ref in tree

This patches adds the basic architecture to adopt api ref in tree
and a new tox command to build api ref for testing.

A new api ref gate job will be added to verify the format which
will be tracked by another commit against project-config.

blueprint: api-ref-in-tree

Change-Id: Idffbdc9325868a3dcd6a9bd7f90f030fd52ee0dc