Updated openstack/openstack Project: openstack-infra/tripleo-ci  8da89169bc49f45043b5b34d0cad59e03b763c9b
Change DLRN promote method

Currently we rely on being able to ssh directly to the DLRN server
in order to promote our current-tripleo symlink. However, RDO is
moving that server to a physical machine behind the ci.centos.org
firewall.

Fortunately, the ci.centos.org jenkins slaves are also on the right
side of this firewall. I have setup a simple job there that takes a
DLRN hash via parameter in HTTP POST, and runs the promote script
for TripleO. This POST is authenticated via a token that will need
to be put in the env on the mirror server.

Change-Id: I7e055c311872fabffc37b5426ca64023c57e6b0a