Updated openstack/openstack Project: openstack-infra/project-config  c4512649ca24acf3cd60406fb7675a4d3d4c985e
Add release announcement jobs for major OSA repos

In order to facilitate the execution of the release announcements for
OpenStack-Ansible repositories (which don't publish tarballs), this
patch sets up the announce-release post job for a selective set of
repositories.

Also included is a small fix to the experimental jobs for the os_aodh
repository.

Change-Id: Ia56d236d3a952f9370a0c23622ec08d62f0e287b