Updated openstack/openstack Project: openstack/ironic  fb4e545e21df8816bec10b604c4c920260e96e79
Use messaging notifications transport instead of default

The usage of oslo_messaging.get_transport is not meant for
notifications; And oslo_messaging.get_notification_transport is
favored for those means. So this change introduces the usage of that
function.

If the settings for the notifications are not set with the
configuration that's under the oslo_messaging_notifications group,
this will fall back to the old settings which are under the DEFAULT
group; just like this used to work.

Change-Id: Ic72fd658fe9b35d39cc3d7da2e4b9fe3454edfba