Updated openstack/openstack Project: openstack/tripleo-heat-templates  7903fc743bd3e445ef48c2d08a4bd3115f825e0b
Cleanup hieradata to reduce Puppet warnings

Some puppet parameters were deprecated, some of them removed.
This patch reduce the number of warnings to a few, and the rest of
warnings are bugs that are in progress by Puppet OpenStack team.

This patch is mostly some cleanup so we don't have useless warnings in
Puppet catalog.

Changes:
* Update Ceilometer auth params
* Update Neutron auth params
* Update Heat auth params
* Update Swift hash suffix param
* Remove neutron::server::notifications::nova_url, useless.

Change-Id: Ie32681a1fe32735f70ba372630da09f91227298c