Updated openstack/openstack Project: openstack/manila  d6f27aa9ffa1d7b4f1f5956d59a10191be4c5afe
Document instructions for documentation

A lot of things have been changing with OpenStack's
documentation. As a community, we could do a better
job at catching up with these changes and
documenting our project well enough for the
benefit of all stakeholders.

Add instructions on where to document what and
provide tips to manila developers on documentation.

Change-Id: I660bfd5ac3d61c253e1e7b4b02c51424a03fa0a0
Partially-implements: bp improve-manila-developer-docs