Updated openstack/openstack Project: openstack/nova  062746d07cce41f2171493df31703a9fca200e07
Fix test_age_and_verify_swap_images

This test mocked os.remove, but the faked implementation recursively
called os.remove if given a path outside the base directory. This
resulted in infinite recursion. We fix this.

The test did not ensure that info files were created under the same
fake base directory as cache entries. This did not cause a failure in
the current code, but a subsequent change trips this.

We update the faking of exists to only return True for files in
expected_exist, which is also tripped in a subsequent change.

We also make a couple of trivial cleanups in related code.

Change-Id: I4f3685bb525ef7b601be3bb7a28983dd4085e68f