Updated openstack/openstack Project: openstack/tempest  2e397f7507ef299c45e209d403370e6f2f74dddd
Add admin role on domain for v3

In case of identity v3, isolated admin users get the admin role
assigned on the project, but not on the domain.
If policy.v3cloudsample.json is in use, the admin role is required
on the domain for several admin actions to be performed.
Extending the CredsClient to support adding role on domain.

This patch changes the creds client for v3, and as there was
no unit test coverage for dynamic creds provider using creds
client on v3, adding the tests.

Change-Id: Iaea458fc8a24f6831476c9ec37cb11d253fcd0ec
Closes-bug: #1494291