Updated openstack/openstack Project: openstack/glance  9d7efa916e2a77c878e02a19e2239a1db9d48dcd
Remove TODOs from deprecated "sign-the-hash"

Most of the code for the deprecated "sign-the-hash" approach
was removed in commit I8862f6c94538dd818c7360ba287e14c1264ff20f.

This patch removes the remainder of the code that was deprecated
for the "sign-the-hash" approach, including the optional
RSA-PSS properties, which are not used in the "sign-the-data"
approach (see the documentation impact section of the spec [1]).

[1] http://specs.openstack.org/openstack/glance-specs/specs/mitaka/approved/image-signing-and-verification-support.html#documentation-impact

Change-Id: Ifdafbed0e3413cc5732a22cc5c9e5adc87487e80