Updated openstack/openstack Project: openstack/oslo.config  1b3af11d9cf4d658015442c92906f4e3d433c594
Return [] for .config_dirs when config files are not parsed

Currently, we raise NoSuchOptError in case config files were not parsed
yet, which does not seem like correct behaviour for the property (since
it is never expected to be an option defined).

Instead of raising, let's return an empty list.

Related-Bug: #1587359
Change-Id: Ia66e9d26b011c0f7bd3d1383d9282554cf300f29