Updated openstack/openstack Project: openstack/ironic  230bca819babc161bd97ea18772cbbada136524f
Remove links to github for OpenStack things

Github is not the official reference for OpenStack code repositories,
and as such should not be used in documentation.

Update links that are meant to be links to a project as a whole with
links to the project's documentation.

Update links directly to code to git.openstack.org.

Remove link to tripleo-incubator script for DIB dependencies, as it does
way more than we need to, and the dependencies needed differ between DIB
elements. Instead we leave this as an exercise to the user (this is
documented in the DIB docs).

Change-Id: Iee09b6b86628e46dc5e9436bd0f98a40c50244b7