Updated openstack/openstack Project: openstack-infra/project-config  a6a008b6842ed310334013ffc87bff230f1f89d2
Remove integrated jobs from neutron-fwaas project config

The neutron integrated jobs (gate-tempest-dsvm-neutron-full and
gate-grenade-dsvm-neutron) are still configured to run with fwaas
(and lbaas) extension, however the decoupling is  ongoing (as
shown in [1]).

Running these jobs on fwaas changes is not only wasteful but it
will eventually not provide any test coverage, so they can be
removed.

[1] Ic60cd1fa90c19dfac00be583e2ddc5633dbb68a3

Change-Id: Id7c6af8d52a779ca0861970333d4c45d89aee566