Updated openstack/openstack Project: openstack/tripleo-heat-templates  c79c261547b1ec8f32c088ab5d4d71af5938c20b
Enable proxy header handling for nova

Nova is using http_proxy_to_wsgi middleware[1][2]. This parses the
headers provided by the proxy, and helps us properly use TLS for
keystone discovery. There was an option introduced in this middleware
to have it disabled by default, and this change enables it.

[1] Ia78f73e96585ab33a379a0b0be6d9682f7fbd810
[2] I808469f24066d382decf55b9dad5312d6e068da7

Change-Id: I3918f24c0c87cb626a28645b46e3df6360d5f924