Updated openstack/openstack Project: openstack/cinder  19fc7db0d913aa0bda4ed16c6dc74751d3b15321
Huawei replication should use auto recovery policy

The recovery policy of replication in huawei driver currently is
"Manual". In this case, when a replication recover from a fault, the
replication on storage becomes "to be restored" state. It requires
the administrator to manually synchronize to recover on storage.
Now we change the recovery policy to "Auto". When the stuff happens,
the replication will be executed automatically synchronized.

Change-Id: I78a158454ed613ba9ff8da0169af82600825247e
Closes-Bug: #1586282