Updated openstack/openstack Project: openstack/neutron  e954ceea14e3979613c75760b06b6f75e8ffbe99
Guard against config_dirs not defined on ConfigOpts

Turned out that if the code extracts config_dirs value from ConfigOpts
objects before config files are parsed, then oslo.config will raise
NoSuchOptError exception.

This is not a usual mode of operation for the code, since main()
function of the process using it is expected to parse CLI and config
files before using it, it may nevertheless happen in some test code.

This patch guards against those exceptions, falling back to
/etc/neutron, as we already do when --config-dir is not specified.

Change-Id: I00cf824baa8580b7aa7ec4518a4741e49c998364
Closes-Bug: #1587359