Updated openstack/openstack Project: openstack/cinder  eb60f3b182dbc11333268edcdcd45d0fe9f674a6
Fix 3PAR cg display name issues

Some special characters and long names cause the creation
and/or updating of consistency groups to fail. There is no longer a
need for saving the display_name in the comments as we already have
the CGs UUID. We will simply remove the display name and the
description.

Closes-Bug: #1573647
Change-Id: I4d7aac74cc1c8a1b7cd5c2a35d40e94d681f7213