Updated openstack/openstack Project: openstack/trove  2f0a8610ca715d297ed30511ec45dd9ae48542e6
Backup to swift using wrong large object type

Trove backups are stored in Swift as Dynamic Large Objects. This is not
ideal, and could cause restore to fail as DLOs may not return all of
their parts. Static Large Objects should be used instead. SLOs perform
checksum and object size validation and can reduce object download times
significantly.

Changes are needed to the Swift storage strategy and taskmanager's
backup delete model to perform actions on SLOs. The corresponding
unittests and fakes are also updated.

A special case occurs when the backup object fits in a single Swift
object segment, in which case it will not be made a SLO.

Change-Id: Ide30a1e7f909a3b7fecb16268367b56468306d42
Closes-bug: 1489997