Updated openstack/openstack Project: openstack-infra/storyboard  176001d6855fe3b1d6def1de38a34d43da7231e9
Updated documentation to make code snippet clearer

Documentation previously showed on screen as two commands on the same
line, which made it appear incorrectly to be one command. I have
changed the code so it now shows each command on a separate line
and in code documentation format as opposed to standard text format.

Change-Id: I12b0ca96bb5942e6b1ea331c72a98d3c7f4b6580