Updated openstack/openstack Project: openstack/nova  0e1093920a17a8576eb55576930cde0b1f08ac72
Remove image cache image verification

There was a substantial amount of code in the image cache devoted to
checksumming images. However, these checksums were never used beyond
the generation of a log message during the image cache manager's
periodic task of the form:

  Corrupt base files: ...

This code was expensive to run, added significant complexity to the
image cache manager, and not actually used by anything. This patch
removes it. The only functional change should be that the above log
message is no longer generated.

We can't remove it completely in this pass because the image cache
may be on shared storage with an old compute host which expects these
files to exist. Consequently we retain the code to delete the image
metadata file when deleting an entry from the cache, and we don't
delete these files automatically when we find them. These vestiges can
be removed in Ocata.

Additional context can be found here:
http://lists.openstack.org/pipermail/openstack-dev/2016-May/095751.html

Change-Id: I3d84117809d5058cb2c779754523913347d5374b