Updated openstack/openstack Project: openstack/requirements  01cafa420ba9aa4b8b04d3e8d1ed5e2672ee5a56
Adding python-oneviewclient to projects.txt

python-oneviewclient is a library used by Ironic to deal with HPE OneView
managed infrastructure. Since it's used by Ironic, requirements shared by
the two projects should be consistent. This patch addes python-oneviewclient
to projects.txt so it can be updated when the global-requirements changes for
projects.

Depends-On: Ia32d66aa7af07f69e1dfd2c28bc8ca96c0832376
Change-Id: I721f2b527cf8249364a7bb670e4a58d43b4f1354