Updated openstack/openstack Project: openstack-infra/gear  3fb6b1c5a1519a47221ec60c145b7a5bdfe5a3a3
Do not change object size when iterating

In `gear.Server._lostConnection()` the jobs list pulls .values() from
`conn.related_jobs`, in `gear.server._removeJob()` it is possible to
change the conn.related_jobs dictionary which will cause a
`RuntimeError: dictionary changed size during iteration` exception to
be raised in python 3 since .values() is an iterator instead of a
rendered list() like in python 2. This RuntimeError is eliminated
by forcing the python 2 behavior by wrapping the `.values()` call
to be wrapped into an explicit list.

Change-Id: I730473f3200a427cce774df534f1ff4958866dac