Updated openstack/openstack Project: openstack/nova  089c6cbf2d47fdf000e46f16f9ad102f185f3030
pci: make sure device relationships are kept in memory

`pci_devs` attribute of PciDevTracker class is the in-memory
"master copy" of all
devices on each compute host, and all data changes that happen when
claiming/allocating/freeing
devices HAVE TO be made against instances contained in `pci_devs`
list, because they are periodically flushed to the DB when the save()
method is called.

Due to this we need to make sure all the relationships are available to
the code using them (claiming/allocation/freeing methods).

We do this by simply keeping a tree structure by referencing
parent/children from objects themselves. This is done on every update of
the state of PCI devices (on compute service start up, and on every
resource tracker pass), so that this information is always as up to date
as the in memory view of devices.

This change adds the code to build up the tree, and subsequent changes
will make sure the newly added relationships are used when needed. We
also add 2 non-versioned fields added to PciDevice object to hold the
references.

Co-Authored-By: Sahid Ferdjaoui <sahid.ferdjaoui@redhat.com>

Change-Id: Id6868b7839efb2cd53f5f7aaac2c55d169356ce4
Partial-bug: #1565785