Updated openstack/openstack Project: openstack/nova  4e767d8fc7c7b762fdf58af4f29b496d55f1f80c
pci: Make sure PF is 'available' when last VF is freed

We were adding the PF back to the pools but not setting the status
properly.

The reason this was not caught by tests is because tests were broken as
well (we were using assertTrue instead of assertEqual which always
passes).

Change-Id: I62d4a810d8d7c4453865db0290029c269225c139
Closes-bug: #1567549