Updated openstack/openstack Project: openstack/nova-specs  4ed51d077ed38a7592c1da0f5ab5865d20fa8560
do not allow specs with zwsp characters

For some reason a number of specs (and docs) have zwsp characters
inserted in URLs. This causes crud in the reviews, and really
confusing errors around line lengths for people (we count them, but
they aren't seen in their editor).

This removes them and makes a test to make sure they never come back.

Change-Id: Ib6e94725e2716dbb2ff86f272fbe4a6987714444