Updated openstack/openstack Project: openstack-infra/infra-specs  1b4ed2fa6889558310b23681f33cad98aed7dc37
Further updates to defaulting to alt election repo

We settled on publishing the election repo to the existing
governance.openstack.org rather than creating a new vhost. This
was prevoiusly an alternative and was switched to be the default
implementation of the spec. However parts missed that switch over.
Fix em' up.

Change-Id: I85db29f35030b3b7aa47699fe962ae79b2793b38