Updated openstack/openstack Project: openstack/nova  bfc5285b4684f6e459e6232ede58365fd9d50e34
libvirt: Skip CPU compatibility check for emulated guests

Comparing host to guest CPU model for emulated guests (<domain
type='qemu'>) should not matter -- in this mode the CPU is fully
emulated in software and no hardware acceleration, like KVM, is
involved.

So, skip the CPU compatibility check for the QEMU domain type,
and retain it for KVM guests.

Co-Authored-By: Daniel P. Berrange <berrange@redhat.com>
Change-Id: I8b782678a6dfa7bcfc203381037e1df4475c3194